fix: Fix client types & move BoardProps to React package (#752)
* refactor(packages): Expose BoardProps type via 'boardgame.io/react'

BoardProps depends on some types from `@types/react`, which won’t be 
installed for TS users not using React, so exporting it from the 
top-level `'boardgame.io'` module causes errors for those users.

* refactor(client): Structure debug types to output a .d.ts file

This places the Svelte Debug component type definition in an index.ts 
alongside the index.js that re-exports the Svelte component.

* Revert "refactor(client): Structure debug types to output a .d.ts file"

This reverts commit e8f07355b7f0d910b081fa79c4518efe2bf2a7d6.

* refactor(client): Add svelte types to tsconfig, typing `.svelte` files

* refactor(packages): Convert debug package to TS

* feat(types): Hoist ambient  declaration in main types file

* chore(deps): Make Svelte a dependency

For TS users, Svelte needs to be installed so its ambient types can be 
used.

* build(tsconfig): Remove `types` field

Now the Svelte types are imported in `src/types.ts`, we don’t need to 
tell `tsc` about them in `tsconfig`.
