+++ packages/react.ts
@@ -6,7 +6,7 @@
  * https://opensource.org/licenses/MIT.
  */
 
-import { Client } from '../src/client/react';
+import { Client, BoardProps } from '../src/client/react';
 import Lobby from '../src/lobby/react';
 
-export { Client, Lobby };
+export { Client, BoardProps, Lobby };
