+++ rollup.config.js
@@ -69,12 +69,6 @@ export default [
   {
     input: subpackages.reduce((obj, name) => {
       obj[name] = `packages/${name}.ts`;
-
-      // The debug package can't be converted to TS
-      // yet due to the svelte import.
-      if (name == 'debug') {
-        obj[name] = 'packages/debug.js';
-      }
       return obj;
     }, {}),
     external,
