--- src/client/debug/index.js
@@ -1,2 +1,0 @@
-import Debug from './Debug.svelte';
-export default Debug;
