+++ src/types.ts
@@ -1,6 +1,7 @@
 import { Object } from 'ts-toolbelt';
 import Koa from 'koa';
 import { Store as ReduxStore } from 'redux';
+import 'svelte';
 import * as ActionCreators from './core/action-creators';
 import { Flow } from './core/flow';
 import { CreateGameReducer } from './core/reducer';
@@ -10,7 +11,6 @@ import { EventsAPI } from './plugins/plugin-events';
 import { RandomAPI } from './plugins/plugin-random';
 
 export { StorageAPI };
-export { BoardProps } from './client/react';
 
 export type AnyFn = (...args: any[]) => any;
